Fix #82: split routing-tuner.ts by concern
routing-tuner.ts mixed data access, text extraction, lock acquisition,
and orchestration in one ~782-line file. Split into a routing-tuner/
directory:

- types.ts            shared interfaces
- text-extraction.ts  keyword/stop-word extraction (pure)
- queries.ts          read-only data access
- lock.ts             advisory lock + tuning_run record lifecycle
- orchestrator.ts     the tune() pipeline + step helpers

routing-tuner.ts becomes a thin barrel re-exporting the public API, so
no importers change. Behavior is unchanged; the capability-lookup,
existing-utterance, and task-count queries are extracted into named
query helpers.

Audit ref: M28.
